feat: fp-1737 move source css
With upcoming fp-1737, moving source stylesheets could make sense.
