+++ taccsite_styles/.postcssrc.yml
@@ -5,7 +5,7 @@ plugins:
     # In any source file, every @import path, not prepended with "./" nor "../",
     # is relative to only one of these directories in this order.
     path:
-      - 'taccsite_cms/static/site_cms/css/src' # Core-CMS CSS source files
+      - 'taccsite_styles/src' # Core-CMS CSS source files
       - 'node_modules/@tacc/core-styles/src/lib' # Core-Styles CSS source files
   postcss-env-function:
     importFrom:
